requirements.txt
# wsgi server
gunicorn==20.1.0

# django
django==2.2.22
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.1
djangorestframework==3.12.4
django-filter==2.4.0

# db
mysqlclient==2.0.3

# celery
celery==5.0.5
flower==0.9.5
django-celery-beat==2.2.0

# redis
django-redis==4.12.1

# shell
ipython==7.22.0

# images
Pillow==8.1.2

# search
django-haystack==3.0

# forum
django-machina==1.1.4

# tree
django-mptt==0.12.0

# widgets
django-widget-tweaks==1.4.8

markdown2==2.4.0

# api doc
drf-yasg==1.20.0

# CORS
django-cors-headers==3.7.0

